Web App/requirements.txt
flask==2.0.1
numpy==1.21.0
opencv-python==4.5.3.56
tensorflow==2.6.0
requests==2.26.0
Pillow==8.3.1
markdown==3.4.1 